lendit-backend/ml/requirements.txt
pandas
scikit-learn
joblib
sqlalchemy
pymysql
PyYAML
